api_yamdb/requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
Django==2.2.6
django-filter==2.4.0
django-rest-authtoken==2.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==2.10
importlib-metadata==4.6.1
iniconfig==1.1.1
numpy==1.21.0
packaging==21.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.1
requests==2.25.1
simplejwt==2.0.1
sqlparse==0.4.1
toml==0.10.2
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
